+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## UNRELEASED
 
+### Added
+
+- You can show/hide each testcase, and only run the non-hidden testcases.
+- Add a drop-down menu for each testcase. You can delete a testcase / run a single testcase from it.
+- Add a drop-down menu for testcases. You can remove empty, remove all, hide AC, show all, hide all, inverse hide/show from it.
+
+### Changed
+
+- The "Hide AC" button is removed. (But a similiar action is added in the "More" menu of testcases.)
+
 ## v6.0
 
 It's the first version after the versioning rules changed, so there's no change log before v6.0.